feat: add sh field in tasks

+++ notion-manager-api/migrations/20250806180916_add_sh_column_in_tasks_table.sql
@@ -1,6 +1,6 @@
 -- +goose Up
 -- +goose StatementBegin
-ALTER TABLE tasks ADD COLUMN sh DOUBLE NOT NULL DEFAULT 0;
+ALTER TABLE tasks ADD COLUMN sh FLOAT NOT NULL DEFAULT 0;
 -- +goose StatementEnd
 
 -- +goose Down
